Modified the code to make use of the new functions for the noise.
One aspect still to be tested is the updated version of the 'step' method, trying to deal with some of the issues encountered when using autograd/backward, trying to unroll derivatives for the entire past. The current tentative solution is an attempt to remove dependences across different time stamps that require computational graphs to be retained.