Finally found the problem, though I'm not sure I know how to fix it.
The trick in the old piece of code is here: "y[2, 0] = y[1, 1]                           # manually assign the acceleration as observed by the agent"
remove this and we get the same oscillations.

We need to find a way to automatically update y[2, 0]. The solution might be to consider 2x2 matrices but with one order of gen. coord., to see if the data generated fills in the gap.